requirements.txt
appdirs==1.4.3
attrs==19.3.0
autopep8==1.5
bcrypt==3.1.7
beautifulsoup4==4.9.0
bs4==0.0.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
click-plugins==1.1.1
colorama==0.4.3
cryptography==2.9
cssselect==1.1.0
dnspython==1.16.0
entrypoints==0.3
fake-useragent==0.1.11
flake8==3.7.9
flake8-docstrings==1.5.0
flake8-double-quotes==0.0.1
flake8-import-order==0.18.1
flake8-polyfill==1.0.2
future==0.18.2
idna==2.9
importlib-metadata==1.6.0
ipwhois==1.1.0
isort==4.3.21
lxml==4.5.0
mccabe==0.6.1
more-itertools==8.2.0
msgpack==1.0.0
netifaces==0.10.9
packaging==20.3
paramiko==2.7.1
parse==1.15.0
pep8-naming==0.10.0
Pillow==7.1.1
pluggy==0.13.1
py==1.8.1
pycodestyle==2.5.0
pycparser==2.20
pydocstyle==5.0.2
pyee==7.0.1
pyflakes==2.1.1
pymetasploit3==1.0.2
PyMySQL==0.9.3
PyNaCl==1.3.0
pyparsing==2.4.7
pyppeteer==0.0.25
pyquery==1.4.1
pytest==5.4.1
python-dateutil==2.8.1
python-dotenv==0.12.0
python-nmap==0.6.1
python-whois==0.7.2
requests==2.23.0
requests-html==0.10.0
selenium==3.141.0
shodan==1.23.0
six==1.14.0
snowballstemmer==2.0.0
soupsieve==2.0
SQLAlchemy==1.3.16
tabulate==0.8.7
tqdm==4.45.0
ujson==2.0.3
urllib3==1.25.8
w3lib==1.21.0
wcwidth==0.1.9
websockets==8.1
XlsxWriter==1.2.8
xmltodict==0.12.0
zipp==3.1.0
